# PERF: Series.corr/Series.cov for EA dtypes