- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.1.0.rst` file if fixing a bug or adding a new feature.

```
import pandas as pd
import numpy as np

N = 1_000_000

ser1 = pd.Series(np.random.randn(N), dtype="Float64")
ser2 = pd.Series(np.random.randn(N), dtype="Float64")

%timeit ser1.corr(ser2)

# 849 ms ± 56.1 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)    -> main
# 33.7 ms ± 320 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR

%timeit ser1.cov(ser2)

# 739 ms ± 15.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)    -> main
# 32.9 ms ± 713 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)  -> PR
```


Also, aligns the error raised for non-numeric data with that of `DataFrame.corr` and `DataFrame.cov`.